+++ x/btcstaking/types/genesis.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"slices"
 	"sort"
 
 	bbn "github.com/babylonlabs-io/babylon/v4/types"
@@ -85,14 +84,6 @@ func (gs GenesisState) Validate() error {
 		return err
 	}
 
-	if err := gs.validateAllowedMultiStakingTxHashes(); err != nil {
-		return err
-	}
-
-	if err := gs.validateAllowedStakingTxHashes(); err != nil {
-		return err
-	}
-
 	if err := gs.validateFpBbnAddr(gs.FinalityProviders); err != nil {
 		return err
 	}
@@ -123,33 +114,6 @@ func (h AllowedStakingTxHashStr) Validate() error {
 	// NewHash validates hash size
 	if _, err := chainhash.NewHash(bz); err != nil {
 		return err
-	}
-	return nil
-}
-
-// validateAllowedStakingTxHashes validates there're no duplicate entries
-// and the hash has the corresponding size
-func (gs GenesisState) validateAllowedStakingTxHashes() error {
-	return validateTxHashes(gs.AllowedStakingTxHashes)
-}
-
-// validateAllowedMultiStakingTxHashes validates there're no duplicate entries
-// and the hash has the corresponding size
-func (gs GenesisState) validateAllowedMultiStakingTxHashes() error {
-	return validateTxHashes(gs.AllowedMultiStakingTxHashes)
-}
-
-func validateTxHashes(hashes []string) error {
-	seen := make(map[string]bool)
-	for _, hStr := range hashes {
-		if _, exists := seen[hStr]; exists {
-			return fmt.Errorf("duplicate staking tx hash: %s", hStr)
-		}
-		seen[hStr] = true
-		h := AllowedStakingTxHashStr(hStr)
-		if err := h.Validate(); err != nil {
-			return err
-		}
 	}
 	return nil
 }
@@ -336,9 +300,6 @@ func SortData(gs *GenesisState) {
 	sort.Slice(gs.ConsumerEvents, func(i, j int) bool {
 		return gs.ConsumerEvents[i].ConsumerId < gs.ConsumerEvents[j].ConsumerId
 	})
-
-	slices.Sort(gs.AllowedStakingTxHashes)
-	slices.Sort(gs.AllowedMultiStakingTxHashes)
 
 	sort.Slice(gs.FpBbnAddr, func(i, j int) bool {
 		return gs.FpBbnAddr[i] < gs.FpBbnAddr[j]